+++ csrc/include/natten/cuda/fna_blackwell/fna_forward.cuh
@@ -76,17 +76,17 @@ struct KernelForward {
   using ProblemShapeType =
       cute::tuple<int, int, int, cute::tuple<cute::tuple<int, int>, int>>;
 
-  using StrideQ =
-      cute::tuple<int, _1, cute::tuple<cute::tuple<int, int>, int>>; // Q D (H_G
-                                                                     // H_R B)
-  using StrideK =
-      cute::tuple<int, _1, cute::tuple<cute::tuple<_0, int>, int>>; // K D (H_G
-                                                                    // H_R B)
+  using StrideQ = cute::
+      tuple<int, _1, cute::tuple<cute::tuple<int, int>, int64_t>>; // Q D (H_G
+                                                                   // H_R B)
+  using StrideK = cute::
+      tuple<int, _1, cute::tuple<cute::tuple<_0, int>, int64_t>>; // K D (H_G
+                                                                  // H_R B)
   using StrideV = StrideK;
   using StrideO = StrideQ;
   using StrideLSE =
-      cute::tuple<int, cute::tuple<cute::tuple<_1, int>, int>>; // Q   (H_G H_R
-                                                                // B)
+      cute::tuple<int, cute::tuple<cute::tuple<_1, int>, int64_t>>; // Q   (H_G
+                                                                    // H_R B)
 
   using TileScheduler = std::conditional_t<
       kIsPersistent,
@@ -174,13 +174,24 @@ struct KernelForward {
     // shape: (batch, seqlen, heads, dim)
     // stride: (dim*heads*seqlen, dim*heads, dim, 1)
     auto stride_Q = make_stride(
-        H * D, _1{}, make_stride(make_stride(D, H_Q * D), H * D * SQ));
+        H * D,
+        _1{},
+        make_stride(
+            make_stride(D, H_Q * D),
+            static_cast<int64_t>(H * D) * static_cast<int64_t>(SQ)));
     auto stride_O = stride_Q;
     auto stride_K = make_stride(
-        H_K * D, _1{}, make_stride(make_stride(_0{}, D), H_K * D * SK));
+        H_K * D,
+        _1{},
+        make_stride(
+            make_stride(_0{}, D),
+            static_cast<int64_t>(H_K * D) * static_cast<int64_t>(SK)));
     auto stride_V = stride_K;
-    auto stride_LSE =
-        make_stride(H, make_stride(make_stride(_1{}, H_Q), SQ * H));
+    auto stride_LSE = make_stride(
+        H,
+        make_stride(
+            make_stride(_1{}, H_Q),
+            static_cast<int64_t>(SQ) * static_cast<int64_t>(H)));
 
     cutlass::KernelHardwareInfo hw_info;
     hw_info.device_id = device_id;
